[DEPYLER-0277] Fix Optional None Return Bug
Fixed return statement codegen generating Ok(()) instead of Ok(None) for Optional return types.

Root Cause:
- codegen_return_stmt() had logic for `is_optional_return && !is_none_literal` (wrap in Some())
- But no explicit branch for `is_optional_return && is_none_literal`
- None literals fell through to general case, generating Ok(()) instead of Ok(None)
- Affected all functions with Optional[T] return type annotations

Fix Applied:
- Added explicit branches in stmt_gen.rs:204-210, 223-229
- Case 1 (Result<Option<T>>): return Ok(None) instead of Ok(())
- Case 2 (Option<T>): return None instead of ()
- Handles both early returns and final returns

Changes:
- crates/depyler-core/src/rust_gen/stmt_gen.rs: Added is_none_literal branches
- docs/bugs/DEPYLER-0277-NONE-RETURN-BUG.md: Comprehensive bug analysis
- docs/bugs/DEPYLER-0278-FNV-DEPENDENCY-MISSING.md: Documented separate fnv issue
- examples/showcase/*.rs: Re-transpiled all showcase examples with fix
- CHANGELOG.md: Added v3.19.2 entry documenting fix

Verification:
- process_config.py: transpiles and compiles with zero errors/warnings ✅
- annotated_example.py: generates Ok(None) on line 79 (still has DEPYLER-0278 fnv issue)
- All other showcase examples still compile

Extreme TDD Cycle:
- RED: process_config.rs failed with "expected Option<String>, found ()"
- GREEN: Added explicit is_none_literal handling for Optional types
- REFACTOR: Validated on process_config and annotated_example

Impact:
- Fixes P1 blocking bug affecting all Optional return types
- Enables correct compilation of functions returning None for Optional types

+++ examples/showcase/binary_search.rs
@@ -1,7 +1,7 @@
 #[doc = "Find target in sorted array, return -1 if not found."]
 pub fn binary_search(arr: &Vec<i32>, target: i32) -> Result<i32, Box<dyn std::error::Error>> {
     let mut left: i32 = 0;
-    let _cse_temp_0 = arr.len();
+    let _cse_temp_0 = arr.len() as i32;
     let mut right: i32 = _cse_temp_0 - 1;
     while left <= right {
         let mid: i32 = {
